This is a few files to be able to point the ship in the correct direction. Usage: run orientvector(vector,pitchP,pitchD, yawP, yawD). run exampleOrientationScript
You can do so by runnig orientvector.ks with parameters for gains. This
includes the Proportional and Deriative gain. A large part of this code
was written by slushat.

Note: I tested this function out and it works, however, it is very
sensitive to the gains that you give it. We will need to work out all the
gains for each stage and all possible scenarios.

This still needs to be optimized to be more fuel efficient. Timing and
duration of the RCS thruster (if used).

EDIT: Copied the wrong example file.